refactor: use check sub tree active menthod from docs-tools
