+++ packages/apps/docs/src/pages/blogchain/2020/index.tsx
@@ -6,12 +6,9 @@ import {
   contentClassVariants,
 } from '@/components/Layout/components/articleStyles.css';
 import { getInitBlogPosts } from '@/hooks/useGetBlogs/utils';
-import {
-  checkSubTreeForActive,
-  getPathName,
-} from '@/utils/staticGeneration/checkSubTreeForActive.mjs';
 import { getData } from '@/utils/staticGeneration/getData.mjs';
 import type { IMenuData, IPageProps } from '@kadena/docs-tools';
+import { checkSubTreeForActive, getPathName } from '@kadena/docs-tools';
 import classNames from 'classnames';
 import type { GetStaticProps } from 'next';
 import type { FC } from 'react';
@@ -47,7 +44,7 @@ export const getStaticProps: GetStaticProps = async () => {
 
   return {
     props: {
-      leftMenuTree: checkSubTreeForActive(getPathName(__filename)),
+      leftMenuTree: await checkSubTreeForActive(getPathName(__filename)),
       posts,
       frontmatter: {
         title: 'BlogChain 2020',
