+++ packages/apps/docs/src/pages/pact/api/index.tsx
@@ -1,11 +1,8 @@
 import apiSpecs from '@/_generated/specs/pact.openapi.json';
 import { options } from '@/components/Layout/Redocly/Redocly';
 import { Specs } from '@/components/Specs/Specs';
-import {
-  checkSubTreeForActive,
-  getPathName,
-} from '@/utils/staticGeneration/checkSubTreeForActive.mjs';
 import type { ILayout } from '@kadena/docs-tools';
+import { checkSubTreeForActive, getPathName } from '@kadena/docs-tools';
 import { Heading } from '@kadena/react-ui';
 import type { GetStaticProps } from 'next';
 import type { OpenAPIV3 } from 'openapi-types';
@@ -25,7 +22,7 @@ const Home: FC<ILayout> = () => {
 export const getStaticProps: GetStaticProps = async () => {
   return {
     props: {
-      leftMenuTree: checkSubTreeForActive(getPathName(__filename)),
+      leftMenuTree: await checkSubTreeForActive(getPathName(__filename)),
       frontmatter: {
         title: 'Pact OpenAPI',
         menu: 'Docs',
