+++ packages/apps/docs/src/pages/pact/index.tsx
@@ -11,11 +11,8 @@ import {
 import MostPopular from '@/components/MostPopular/MostPopular';
 import { getBlogPosts } from '@/utils/getBlogPosts';
 import getMostPopularPages from '@/utils/getMostPopularPages';
-import {
-  checkSubTreeForActive,
-  getPathName,
-} from '@/utils/staticGeneration/checkSubTreeForActive.mjs';
 import type { IMenuData } from '@kadena/docs-tools';
+import { checkSubTreeForActive, getPathName } from '@kadena/docs-tools';
 import {
   Box,
   Button,
@@ -157,7 +154,7 @@ export const getStaticProps: GetStaticProps = async () => {
     props: {
       popularPages: mostPopularPages,
       blogPosts,
-      leftMenuTree: checkSubTreeForActive(getPathName(__filename)),
+      leftMenuTree: await checkSubTreeForActive(getPathName(__filename)),
       frontmatter: {
         title: 'Learn Pact',
         subTitle: 'The human-readable smart contract language',
