+++ packages/apps/docs/src/pages/tags/index.tsx
@@ -7,11 +7,8 @@ import {
 import { TagList } from '@/components/TagList/TagList';
 import { TagListItem } from '@/components/TagList/TagListItem';
 import { getAllBlogTags } from '@/utils/getAllBlogTags';
-import {
-  checkSubTreeForActive,
-  getPathName,
-} from '@/utils/staticGeneration/checkSubTreeForActive.mjs';
 import type { IPageProps, ITag } from '@kadena/docs-tools';
+import { checkSubTreeForActive, getPathName } from '@kadena/docs-tools';
 import { Stack } from '@kadena/react-ui';
 import classNames from 'classnames';
 import type { GetStaticProps } from 'next';
@@ -46,7 +43,7 @@ export const getStaticProps: GetStaticProps = async () => {
   const tags = getAllBlogTags();
   return {
     props: {
-      leftMenuTree: checkSubTreeForActive(getPathName(__filename)),
+      leftMenuTree: await checkSubTreeForActive(getPathName(__filename)),
       tags,
       frontmatter: {
         title: 'Tags',
